+++ Map.cpp
@@ -18,15 +18,15 @@ int Map::getCellSize() const {
 }
 
 void Map::drawGrid(sf::RenderWindow& window) {
-    sf::RectangleShape line(sf::Vector2f(getLength(), 1)); // Horizontal line
+    sf::RectangleShape line(sf::Vector2f(static_cast<float>(getLength()), 1.f)); // Horizontal line
     line.setFillColor(sf::Color::Green);
     for (int i = 0; i <= rows; i++) {
-        line.setPosition(0, i * cellSize);
+        line.setPosition(0, static_cast<float>(i * cellSize));
         window.draw(line);
     }
-    line.setSize(sf::Vector2f(1, getLength())); // Vertical line
+    line.setSize(sf::Vector2f(1.f, static_cast<float>(getLength()))); // Vertical line
     for (int i = 0; i <= cols; i++) {
-        line.setPosition(i * cellSize, 0);
+        line.setPosition(static_cast<float>(i * cellSize), 0);
         window.draw(line);
     }
 }
